% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_X.R
\name{build_X}
\alias{build_X}
\title{Generating Observed Networks - X}
\usage{
build_X(
  synthetic,
  mutuality = 0.5,
  sh_theta = 2,
  sc_theta = 0.5,
  theta = reticulate::py_none(),
  flag_self_reporter = T,
  Q = reticulate::py_none(),
  cutoff_X = F,
  lambda_diff = reticulate::py_none(),
  seed = reticulate::py_none(),
  verbose = T
)
}
\arguments{
\item{synthetic}{A synthetic model}

\item{mutuality}{The mutuality parameter (from 0 to 1)}

\item{sh_theta}{Shape of gamma distribution from which to draw theta. The 'reliability' of nodes is represented by the parameter $theta_{lm}$ and by default are modelled as a gamma function with shape \code{sh_theta} and scale \code{sc_theta}.}

\item{sc_theta}{Scale of gamma distribution from which to draw theta.}

\item{theta}{Custom values of theta, if provided sh_theta and sc_theta will be ignored.}

\item{flag_self_reporter}{Indicates whether a node can only report about their own ties (default is true).}

\item{Q}{Maximum value of X entries. If None, it will use the network's K parameter}

\item{cutoff_X}{Whether to set X as a binary}

\item{lambda_diff}{The difference between each subsequent K}

\item{seed}{Pseudo random generator seed to use}

\item{verbose}{Provides additional details}
}
\value{
A character vector.
}
\description{
Given a network \code{Y}, we can generate N observed adjacency matrices as would
have been reported by reporting nodes $m$ ($m in N$).
}
\examples{

random_net <- GMReciprocity(N=100, M=100, L=1, eta=0.99)
X <- build_X(random_net, flag_self_reporter=TRUE, cutoff_X=FALSE, seed=10L)
dim(X)
}
