% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{check_reporters_mask}
\alias{check_reporters_mask}
\title{Check reporters mask}
\usage{
check_reporters_mask(R)
}
\arguments{
\item{R}{Reporters mask (TRUE/FALSE) indicating whether a node
\emph{CAN} report on a particular tie, with dimensions L x N x N x N.}
}
\value{
TRUE if reporters mask is valid, FALSE otherwise.
}
\description{
Check if reporters mask is valid. That is, if it is an array with four
dimensions, if the second, third, and fourth dimensions have the same
length, and if it is either entirely TRUE/FALSE or entirely 0s/1s.
}
\seealso{
Other vimure: 
\code{\link{get_inferred_model}()},
\code{\link{summary.vimure.model.VimureModel}()},
\code{\link{vimure}()}
}
\concept{vimure}
