% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{DegreeCorrectedSBM}
\alias{DegreeCorrectedSBM}
\title{DegreeCorrectedSBM}
\usage{
DegreeCorrectedSBM(
  N = 100,
  M = 100,
  K = 2,
  L = 1,
  exp_in = 2,
  exp_out = 2.5,
  seed = 10
)
}
\value{
DegreeCorrectedSBM model
}
\description{
DegreeCorrectedSBM
}
