% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{extract_Xavg}
\alias{extract_Xavg}
\title{Extract the Xavg matrix (considering the reporter's mask)}
\usage{
extract_Xavg(synthetic)
}
\arguments{
\item{synthetic}{A synthetic model}
}
\value{
A matrix
}
\description{
Extract the Xavg matrix (considering the reporter's mask)
}
\examples{
random_net <- GMReciprocity(N=100, M=100, L=1, eta=0.99)
X <- build_X(random_net)
Xavg <- extract_Xavg(random_net)
dim(Xavg)
}
