% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{extract_Y}
\alias{extract_Y}
\title{Extract the Y matrix}
\usage{
extract_Y(synthetic)
}
\arguments{
\item{synthetic}{A synthetic model}
}
\value{
A sptensor
}
\description{
Extract the Y matrix
}
\examples{
random_net <- gm_CReciprocity(N=100, M=100, L=1, eta=0.99)
Y <- extract_Y(random_net)
dim(Y)
}
