% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_inferred_model}
\alias{get_inferred_model}
\title{Estimate Y}
\usage{
get_inferred_model(object, method = "rho_max", threshold = NULL)
}
\arguments{
\item{object}{A "vimure" object.}

\item{method}{A character string indicating which method is to be computed.
One of "rho_max" (default), "rho_mean", "fixed_threshold" or "heuristic_threshold".}

\item{threshold}{A threshold to be used when method = "fixed_threshold".}
}
\description{
Estimate Y
}
\details{
Use this function to reconstruct the Y matrix with a fitted vimure model.
It will use \code{model$rho_f} values to extract an estimated Y matrix.
\itemize{
\item{\emph{rho_max}: }{Assign the value of the highest probability}
\item{\emph{rho_mean}: }{Expected value of the multinomial}
\item{\emph{fixed_threshold}: }{Check if the probability is higher than a threshold (Only for 2 categories)}
\item{\emph{heuristic_threshold}: }{Calculate and use the best threshold (Only for 2 categories)}
}
}
