% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{gm_CReciprocity}
\alias{gm_CReciprocity}
\title{Generative Model with Reciprocity (CRep)}
\usage{
gm_CReciprocity(
  N = 100,
  M = 100,
  K = 2,
  L = 1,
  C = 2,
  avg_degree = 10,
  sparsify = TRUE,
  eta = 0.99,
  seed = 10
)
}
\arguments{
\item{N}{Number of nodes}

\item{M}{Number of reporters}

\item{K}{Maximum edge weight in the adjacency matrix. When K=2 (default), the adjacency matrix will contain some Y(ij)=0 and Y(ij)=1.}

\item{L}{Number of layers (it has only been tested for L=1)}

\item{C}{Number of communities}

\item{avg_degree}{Desired average degree for the network. It is not guaranteed that the ultimate network will have that exact average degree value. Try tweaking this parameter if you want to increase or decrease the density of the network.}

\item{sparsify}{If True (default), enforce sparsity.}

\item{eta}{Initial value for the reciprocity coefficient. Eta has to be in [0, 1).}

\item{seed}{Pseudo random generator seed to use}
}
\value{
gm_CReciprocity model
}
\description{
A mathematically principled generative model for capturing both community and reciprocity patterns in directed networks.
For more information about this model, see Safdari, H., Contisciani, M., & De Bacco, C. (2021). \emph{Generative model for reciprocity and community detection in networks}. Physical Review Research, 3(2), 023209.
\href{https://journals.aps.org/prresearch/abstract/10.1103/PhysRevResearch.3.023209}{DOI:10.1103/PhysRevResearch.3.023209}.
}
\examples{
random_net <- gm_CReciprocity(N=100, M=100, L=1, eta=0.5)
Y <- extract_Y(random_net)
dim(Y)
}
\seealso{
Other generative models: 
\code{\link{gm_DegreeCorrectedSBM}()},
\code{\link{gm_StandardSBM}()}
}
\concept{generative models}
