% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{gm_DegreeCorrectedSBM}
\alias{gm_DegreeCorrectedSBM}
\title{Degree-corrected stochastic blockmodel}
\usage{
gm_DegreeCorrectedSBM(
  N = 100,
  M = 100,
  K = 2,
  L = 1,
  C = 2,
  avg_degree = 10,
  exp_in = 2,
  exp_out = 2.5,
  sparsify = TRUE,
  seed = 10
)
}
\arguments{
\item{N}{Number of nodes}

\item{M}{Number of reporters}

\item{K}{Maximum edge weight in the adjacency matrix. When K=2 (default), the adjacency matrix will contain some Y(ij)=0 and Y(ij)=1.}

\item{L}{Number of layers (it has only been tested for L=1)}

\item{C}{Number of communities}

\item{avg_degree}{Desired average degree for the network. It is not guaranteed that the ultimate network will have that exact average degree value. Try tweaking this parameter if you want to increase or decrease the density of the network.}

\item{exp_in}{Exponent power law of in-degree distribution}

\item{exp_out}{Exponent power law of out-degree distribution}

\item{sparsify}{If True (default), enforce sparsity.}

\item{seed}{Pseudo random generator seed to use}
}
\value{
gm_DegreeCorrectedSBM model
}
\description{
A generative model that incorporates heterogeneous vertex degrees into stochastic blockmodels, improving the performance of the models for statistical inference of group structure.
For more information about this model, see Karrer, B., & Newman, M. E. (2011). \emph{Stochastic blockmodels and community structure in networks}. Physical review E, 83(1), 016107.
\href{https://arxiv.org/pdf/1008.3926.pdf}{DOI:10.1103/PhysRevE.83.016107}
}
\examples{
random_net <- gm_DegreeCorrectedSBM(exp_in = 2, exp_out = 2.5)
Y <- random_net$Y$toarray()
dim(Y)
}
\seealso{
Other generative models: 
\code{\link{gm_Multitensor}()},
\code{\link{gm_StandardSBM}()}
}
\concept{generative models}
