% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{gm_StandardSBM}
\alias{gm_StandardSBM}
\title{Standard stochastic blockmodel}
\usage{
gm_StandardSBM(
  N = 100,
  M = 100,
  K = 2,
  L = 1,
  C = 2,
  avg_degree = 10,
  sparsify = TRUE,
  overlapping = 0,
  seed = 10
)
}
\arguments{
\item{N}{Number of nodes}

\item{M}{Number of reporters}

\item{K}{Maximum edge weight in the adjacency matrix. When K=2 (default), the adjacency matrix will contain some Y(ij)=0 and Y(ij)=1.}

\item{L}{Number of layers (it has only been tested for L=1)}

\item{C}{Number of communities}

\item{avg_degree}{Desired average degree for the network. It is not guaranteed that the ultimate network will have that exact average degree value. Try tweaking this parameter if you want to increase or decrease the density of the network.}

\item{sparsify}{If True (default), enforce sparsity.}

\item{overlapping}{Fraction of nodes with mixed membership. It has to be in [0, 1).}

\item{seed}{Pseudo random generator seed to use}
}
\value{
gm_StandardSBM model
}
\description{
A generative graph model which assumes the probability of connecting two nodes in a graph is determined entirely by their block assignments.
For more information about this model, see Holland, P. W., Laskey, K. B., & Leinhardt, S. (1983). \emph{Stochastic blockmodels: First steps. Social networks}, 5(2), 109-137.
\href{https://www.sciencedirect.com/science/article/abs/pii/0378873383900217}{DOI:10.1016/0378-8733(83)90021-7}
}
\examples{
random_net <- gm_StandardSBM(N=100, M=100, L=1)
Y <- extract_Y(random_net)
dim(Y)
}
\seealso{
Other generative models: 
\code{\link{gm_Multitensor}()},
\code{\link{gm_DegreeCorrectedSBM}()}
}
\concept{generative models}
