% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{overall_reciprocity}
\alias{overall_reciprocity}
\title{Reciprocity of networks}
\usage{
overall_reciprocity(X)
}
\arguments{
\item{X}{A matrix}
}
\description{
Calculates the reciprocity of adjancency matrix
}
