% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_graph_from_csv}
\alias{parse_graph_from_csv}
\title{Parses a csv file to vm.io.RealNetwork class}
\usage{
parse_graph_from_csv(
  file,
  weighted = F,
  directed = T,
  ego = "Ego",
  alter = "Alter",
  reporter = "reporter",
  layer = "layer",
  weight = "weight",
  ...
)
}
\arguments{
\item{file}{data.frame representing the edgelist.}

\item{weighted}{Whether to add weights to the adjancency matrix.}

\item{directed}{Whether to create a directed graph.}

\item{ego}{Column name for mapping Ego.}

\item{alter}{Column name for mapping Alter.}

\item{reporter}{Column name for mapping reporter.}

\item{layer}{Column name for mapping layer.}

\item{weight}{Column name for mapping weight.}

\item{...}{Additional args of read.csv function}
}
\value{
vimureP$io$RealNetwork
}
\description{
parse_graph_from_csv creates a vimure.io.RealNetwork from a csv file
Its argument is a csv file with at least Ego and Alter columns, each row defines one edge.
The dataframe can contains three optional columns:
'reporter' for identify who reports each edge, 'layer' for multidimensions networks and 'weight' for
a weighted adjancency matrix
Use the params \code{ego}, \code{alter}, \code{reporter}, \code{weight} and \code{layer} to map the dataframe columns.
}
\seealso{
Other parse data input: 
\code{\link{parse_graph_from_edgelist}()},
\code{\link{parse_graph_from_igraph}()}
}
\concept{parse data input}
