% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_graph_from_igraph}
\alias{parse_graph_from_igraph}
\title{Parses an igraph object to vm.io.RealNetwork class}
\usage{
parse_graph_from_igraph(
  graph,
  directed = T,
  weighted = F,
  weight = "weight",
  reporter = "reporter",
  layer = "layer"
)
}
\arguments{
\item{graph}{An igraph object.}

\item{directed}{Whether to create a directed graph.}

\item{weighted}{Whether to add weights to the adjancency matrix.}

\item{weight}{Edge attribute name for mapping weight.}

\item{reporter}{Edge attribute name for mapping reporter.}

\item{layer}{Edge attribute name for mapping layer.}
}
\value{
vimureP$io$RealNetwork
}
\description{
parse_graph_from_igraph creates a vimure.io.RealNetwork from an igraph object.
The igraph can contains three optional edge attributes:
'reporter' for identify who reports each edge, 'layer' for multidimensions networks and 'weight' for
a weighted adjancency matrix
Use the params \code{reporter}, \code{weight} and \code{layer} to map the edge attribute names.
}
