% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{sample_inferred_model}
\alias{sample_inferred_model}
\title{Sample Y trials from rho distribution
Use this function to sample Y trials with a fitted vimure model.
It will use \code{model.rho_f} as the probabilities of a discrete distribution.}
\usage{
sample_inferred_model(object, N = 1, seed = NULL)
}
\arguments{
\item{object}{A "vimure" object.}

\item{N}{Number of trials.}

\item{seed}{A pseudo generator seed.}
}
\description{
Sample Y trials from rho distribution
Use this function to sample Y trials with a fitted vimure model.
It will use \code{model.rho_f} as the probabilities of a discrete distribution.
}
