% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{StandardSBM}
\alias{StandardSBM}
\title{Standard stochastic block-model synthetic network}
\usage{
StandardSBM(
  N = 100,
  M = 100,
  K = 2,
  L = 1,
  C = 2,
  avg_degree = 10,
  sparsify = TRUE,
  overlapping = 0,
  seed = 10
)
}
\value{
DegreeCorrectedSBM model
}
\description{
Standard stochastic block-model synthetic network
}
