% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{summary.vimure.model.VimureModel}
\alias{summary.vimure.model.VimureModel}
\title{Diagnostics metrics}
\usage{
\method{summary}{vimure.model.VimureModel}(object, net = NULL, ...)
}
\arguments{
\item{object}{A "vimure" object.}

\item{net}{A vm.io.baseNetwork object.}

\item{...}{Additional arguments affecting the summary produced.}
}
\description{
Diagnostics metrics
}
\seealso{
Other vimure: 
\code{\link{get_inferred_model}()},
\code{\link{vimure}()}
}
\concept{vimure}
