% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{synthetic_SBM}
\alias{synthetic_SBM}
\title{Standard stochastic block-model synthetic network}
\usage{
synthetic_SBM(
  N = 100,
  M = 100,
  K = 2,
  L = 1,
  C = 2,
  avg_degree = 10,
  sparsify = TRUE,
  overlapping = 0,
  seed = 10
)
}
\arguments{
\item{N}{Number of nodes}

\item{M}{Number of reporters}

\item{K}{Maximum edge weight in the adjacency matrix. When K=2 (default), the adjacency matrix will contain some Y(ij)=0 and Y(ij)=1.}

\item{L}{Number of layers (it has only been tested for L=1)}

\item{C}{Number of communities}

\item{avg_degree}{Desired average degree for the network. It is not guaranteed that the ultimate network will have that exact average degree value. Try tweaking this parameter if you want to increase or decrease the density of the network.}

\item{sparsify}{If True (default), enforce sparsity.}

\item{overlapping}{n/a}

\item{seed}{Pseudo random generator seed to use}
}
\value{
synthetic_SBM model
}
\description{
More info \href{https://appliednetsci.springeropen.com/track/pdf/10.1007/s41109-019-0170-z.pdf#page=22&zoom=100,148,205}{here}
}
\examples{
random_net <- synthetic_SBM(N=100, M=100, L=1)
Y <- extract_Y(random_net)
dim(Y)
}
