% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{vimure}
\alias{vimure}
\title{ViMuRe}
\usage{
vimure(
  x,
  R = NULL,
  mutuality = TRUE,
  undirected = FALSE,
  theta_prior = c(0.1, 0.1),
  K = NULL,
  lambda_prior = c(10, 10),
  eta_prior = c(0.5, 1),
  rho_prior = NULL,
  seed = NULL,
  etol = 0.1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{An adjancency matrix with dimensions L x N x N x N
or a igraph object.}

\item{R}{Reporters mask (TRUE/FALSE) indicating whether a node
\emph{CAN} report on a particular tie, with dimensions L x N x N x N.
If reporters mask was not informed, the model will assume that
every reporter can report on any tie.}

\item{mutuality}{Whether to use the mutuality parameter.}

\item{undirected}{Whether the network is undirected.}

\item{theta_prior}{Shape and scale hyperparameters for variable theta.}

\item{K}{Value of the maximum entry of the network - i.}

\item{lambda_prior}{Shape and scale hyperparameters for variable lambda.}

\item{eta_prior}{Shape and scale hyperparameters for variable eta.}

\item{rho_prior}{Array with prior values of the rho parameter - if ndarray.}

\item{seed}{Pseudo random generator seed to use.}

\item{etol}{controls when to stop the optimisation algorithm (CAVI).}

\item{verbose}{Provides additional details.}

\item{...}{Additional args of model$fit() method.}
}
\value{
vimure model
}
\description{
Fit a probabilistic generative model to double sampled networks.
It returns reliability parameters for the reporters (theta),
average interactions for the links (lambda) and the estimate of
the true and unknown network (rho).
The inference is performed with a Variational Inference approach.
}
\seealso{
Other vimure: 
\code{\link{check_reporters_mask}()},
\code{\link{get_inferred_model}()},
\code{\link{summary.vimure.model.VimureModel}()}
}
\concept{vimure}
