% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{vm_config}
\alias{vm_config}
\title{Vimure configuration information}
\usage{
vm_config()
}
\value{
List with information on the current configuration of Vimure
You can determine whether Vimure was found using the \code{available}
member,
}
\description{
Vimure configuration information
}
\keyword{internal}
