% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_X.R
\name{extract_Xavg}
\alias{extract_Xavg}
\title{Extract the Xavg matrix}
\usage{
extract_Xavg(synthetic)
}
\arguments{
\item{synthetic}{A synthetic model}
}
\value{
A matrix
}
\description{
Average value of observed network X over reporter mask.
}
\examples{
random_net <- gm_Multitensor(N=100, M=100, L=1, eta=0.99)
X <- build_X(random_net)
Xavg <- extract_Xavg(random_net)
dim(Xavg)
}
