% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{summary.vimure.model.VimureModel}
\alias{summary.vimure.model.VimureModel}
\title{Diagnostics metrics}
\usage{
\method{summary}{vimure.model.VimureModel}(object, net, ...)
}
\arguments{
\item{object}{a "vimure" object}

\item{net}{a "baseNetwork" object}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Diagnostics metrics
}
